修改界面细节，首页、兑换商品详情、

+++ gjtjb/RKWXT/VirtualMall/VirtualGoodsInfo/Model/VirtualGoodsInfoTool.h
@@ -25,7 +25,7 @@
 + (CGFloat)backMoney:(VirtualGoodsInfoModel*)model;
 
 //所需云票
-+ (CGFloat)xnb:(VirtualGoodsInfoModel*)model;
++ (int)xnb:(VirtualGoodsInfoModel*)model;
 
 //价格
 + (CGFloat)goodsPrice:(VirtualGoodsInfoModel*)model;
